create backbone of the logic

+++ model/Planet.cpp
@@ -1,14 +1,14 @@
 #include "Planet.h"
 
-Planet::Planet(string name, int x, int y, int id) : name(name), posX(x), posY(y), sectorID(id){}
+Planet::Planet(string name, int x, int y, size_t id) : name(name), posX(x), posY(y), sectorID(id){}
 
 
 
-void Planet::setSectorID(int id){
+void Planet::setSectorID(size_t id){
     sectorID = id;
 }
 
-int Planet::getSectorID() const {
+size_t Planet::getSectorID() const {
     return sectorID;
 }
